requirements.txt
flask
pytesseract
pillow
numpy
scikit-learn

# https://github.com/UB-Mannheim/tesseract/wiki
